Standardize on 3.10-3.12
Dev-compatibility with a wide range of versions is getting increasingly
hard (and in particular, linting only works on 3.12 due to ast changes).

You can still (probably) test and publish on older versions, but for my
projects this is diminishing returns and the sort of thing I'd need to
ponder on an individual basis.

usort needs LibCST to work on 3.13, and it looks like there's a rust
issue fixed in git so it's just waiting on a release > 1.1.0.